git-svn-id: https://svn.kuali.org/repos/student/enrollment/ks-ap/branches/KSAP-0.8.1-iu@72493 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-ap-framework/src/main/java/org/kuali/student/ap/coursesearch/CourseSearchItem.java
@@ -1,5 +1,6 @@
 package org.kuali.student.ap.coursesearch;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -49,20 +50,20 @@ public interface CourseSearchItem {
 	 * 
 	 * @return The credits minimum value.
 	 */
-	float getCreditMin();
+	BigDecimal getCreditMin();
 
 	/**
 	 * Get the credits maximum value, for use with CreditsFacet.
 	 * 
 	 * @return The credits maximum value.
 	 */
-	float getCreditMax();
+	BigDecimal getCreditMax();
 
     /**
      * Get the multiple values of credits, for use with CreditsFacet
      * @return The array of credit values
      */
-    float[] getMultipleCredits();
+	BigDecimal[] getMultipleCredits();
 
 	/**
 	 * Get the credit type for use with CreditsFacet.
